+++ primitives/src/lib.rs
@@ -38,7 +38,7 @@ sp_api::decl_runtime_apis! {
 	pub trait TimeApi {
 		fn get_shards(peer_id: PeerId) -> Vec<ShardId>;
 		fn get_shard_members(shard_id: ShardId) -> Vec<PeerId>;
-		fn get_shard_threshold(shard_id: ShardId) -> u8;
+		fn get_shard_threshold(shard_id: ShardId) -> u16;
 		fn get_shard_tasks(shard_id: ShardId) -> Vec<TaskExecution>;
 		fn get_task(task_id: TaskId) -> Option<TaskDescriptor>;
 	}
